requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
cachetools==6.2.0
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
db-dtypes==1.4.3
    # via durham-environmental-monitoring (pyproject.toml)
google-api-core==2.26.0
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-auth==2.41.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   google-cloud-storage
google-cloud-appengine-logging==1.6.2
    # via google-cloud-logging
google-cloud-audit-log==0.3.3
    # via google-cloud-logging
google-cloud-bigquery==3.38.0
    # via durham-environmental-monitoring (pyproject.toml)
google-cloud-bigquery-storage==2.33.1
    # via durham-environmental-monitoring (pyproject.toml)
google-cloud-core==2.4.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.12.1
    # via durham-environmental-monitoring (pyproject.toml)
google-cloud-secret-manager==2.24.0
    # via durham-environmental-monitoring (pyproject.toml)
google-cloud-storage==3.4.1
    # via durham-environmental-monitoring (pyproject.toml)
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.14.2
    # via
    #   google-cloud-logging
    #   google-cloud-secret-manager
grpcio==1.75.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.75.1
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via durham-environmental-monitoring (pyproject.toml)
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.7.0
    # via opentelemetry-api
nest-asyncio==1.6.0
    # via durham-environmental-monitoring (pyproject.toml)
numpy==2.3.3
    # via
    #   db-dtypes
    #   pandas
opentelemetry-api==1.37.0
    # via google-cloud-logging
packaging==25.0
    # via
    #   db-dtypes
    #   google-cloud-bigquery
pandas==2.3.3
    # via
    #   durham-environmental-monitoring (pyproject.toml)
    #   db-dtypes
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-bigquery-storage
    #   google-cloud-logging
    #   google-cloud-secret-manager
protobuf==6.33.5
    # via
    #   durham-environmental-monitoring (pyproject.toml)
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-bigquery-storage
    #   google-cloud-logging
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.11
    # via durham-environmental-monitoring (pyproject.toml)
pyarrow==21.0.0
    # via
    #   durham-environmental-monitoring (pyproject.toml)
    #   db-dtypes
pyasn1==0.6.3
    # via
    #   durham-environmental-monitoring (pyproject.toml)
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.0
    # via durham-environmental-monitoring (pyproject.toml)
pydantic-core==2.41.1
    # via pydantic
python-dateutil==2.9.0.post0
    # via
    #   google-cloud-bigquery
    #   pandas
python-dotenv==1.1.1
    # via durham-environmental-monitoring (pyproject.toml)
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via durham-environmental-monitoring (pyproject.toml)
requests==2.33.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.44
    # via durham-environmental-monitoring (pyproject.toml)
tqdm==4.67.1
    # via durham-environmental-monitoring (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   anyio
    #   grpcio
    #   opentelemetry-api
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.6.3
    # via
    #   durham-environmental-monitoring (pyproject.toml)
    #   requests
zipp==3.23.0
    # via importlib-metadata
